+++ runbotics/runbotics-orchestrator-ui/src/src-landing/views/sections/IntegrationSections/IntegrationBackground/index.ts
@@ -1,0 +1,3 @@
+import IntegrationBackground from './IntegrationBackground';
+
+export default IntegrationBackground;
