+++ runbotics/runbotics-orchestrator-ui/src/src-landing/views/sections/IntegrationSections/IntegrationContent/index.ts
@@ -1,0 +1,3 @@
+import IntegrationContent from './IntegrationContent';
+
+export default IntegrationContent;
