+++ runbotics/runbotics-orchestrator-ui/src/src-landing/views/sections/IntegrationSections/IntegrationSection.tsx
@@ -1,0 +1,12 @@
+import { FC } from 'react';
+
+import IntegrationBackground from './IntegrationBackground';
+import IntegrationContent from './IntegrationContent';
+
+const IntegrationSection: FC = () => (
+    <IntegrationBackground>
+        <IntegrationContent />
+    </IntegrationBackground>
+);
+
+export default IntegrationSection;
