+++ runbotics/runbotics-orchestrator-ui/src/src-landing/views/sections/IntegrationSections/index.tsx
@@ -1,3 +1,3 @@
-import IntegrationSection from './IntegrationSections';
+import IntegrationSection from './IntegrationSection';
 
 export default IntegrationSection;
